requirements.txt
certifi==2022.9.14
chardet==3.0.4
idna==2.10
oauthlib==3.2.1
python-dotenv==0.21.0
requests==2.24.0
requests-oauthlib==1.3.0
urllib3==1.25.11